Make examples standalone packages
Motivation:

Standalone packages for examples are an easier on-ramp for newcomers as
they only see the details they care about and it can act as a starting
point which they can develop from.

Modifications:

- Turn each example into a standalone package and add a README to each
- Update the generate and fetch scripts and re-run them

Result:

Better examples
