+++ Examples/echo/Sources/Echo.swift
@@ -17,7 +17,6 @@
 import ArgumentParser
 
 @main
-@available(macOS 15.0, iOS 18.0, watchOS 11.0, tvOS 18.0, visionOS 2.0, *)
 struct Echo: AsyncParsableCommand {
   static let configuration = CommandConfiguration(
     commandName: "echo",
