+++ Examples/hello-world/Sources/HelloWorld.swift
@@ -15,21 +15,12 @@
  */
 
 import ArgumentParser
-import GRPCHTTP2Core
 
-struct ClientArguments: ParsableArguments {
-  @Option(help: "The server's listening port")
-  var port: Int = 1234
-
-  @Option(help: "The number of times to repeat the call")
-  var repetitions: Int = 1
-
-  @Option(help: "Message to send to the server")
-  var message: String
-}
-
-extension ClientArguments {
-  var target: any ResolvableTarget {
-    return .ipv4(host: "127.0.0.1", port: self.port)
-  }
+@main
+struct HelloWorld: AsyncParsableCommand {
+  static let configuration = CommandConfiguration(
+    commandName: "hello-world",
+    abstract: "A multi-tool to run a greeter server and execute RPCs against it.",
+    subcommands: [Serve.self, Greet.self]
+  )
 }
