+++ Examples/hello-world/Sources/Subcommands/Greet.swift
@@ -15,10 +15,9 @@
  */
 
 import ArgumentParser
-import GRPCHTTP2Transport
+import GRPCNIOTransportHTTP2
 import GRPCProtobuf
 
-@available(macOS 15.0, iOS 18.0, watchOS 11.0, tvOS 18.0, visionOS 2.0, *)
 struct Greet: AsyncParsableCommand {
   static let configuration = CommandConfiguration(abstract: "Sends a request to the greeter server")
