+++ Examples/route-guide/Sources/Subcommands/RouteChat.swift
@@ -15,26 +15,18 @@
  */
 
 import ArgumentParser
-import GRPCHTTP2Transport
+import GRPCNIOTransportHTTP2
 
-@available(macOS 15.0, iOS 18.0, watchOS 11.0, tvOS 18.0, visionOS 2.0, *)
-struct GetFeature: AsyncParsableCommand {
-  static let configuration = CommandConfiguration(abstract: "Gets a feature at a given location.")
+struct RouteChat: AsyncParsableCommand {
+  static let configuration = CommandConfiguration(
+    abstract: """
+      Visits a few points and records a note at each, and prints all notes previously recorded at \
+      each point.
+      """
+  )
 
   @Option(help: "The server's listening port")
   var port: Int = 31415
-
-  @Option(
-    name: [.customLong("latitude"), .customLong("lat")],
-    help: "Latitude of the feature to get in E7 format (degrees ⨯ 1e7)"
-  )
-  var latitude: Int32 = 407_838_351
-
-  @Option(
-    name: [.customLong("longitude"), .customLong("lon")],
-    help: "Longitude of the feature to get in E7 format (degrees ⨯ 1e7)"
-  )
-  var longitude: Int32 = -746_143_763
 
   func run() async throws {
     let transport = try HTTP2ClientTransport.Posix(
@@ -50,17 +42,29 @@ struct GetFeature: AsyncParsableCommand {
 
       let routeGuide = Routeguide_RouteGuideClient(wrapping: client)
 
-      let point = Routeguide_Point.with {
-        $0.latitude = self.latitude
-        $0.longitude = self.longitude
-      }
+      try await routeGuide.routeChat { writer in
+        let notes: [(String, (Int32, Int32))] = [
+          ("First message", (0, 0)),
+          ("Second message", (0, 1)),
+          ("Third message", (1, 0)),
+          ("Fourth message", (0, 0)),
+          ("Fifth message", (1, 0)),
+        ]
 
-      let feature = try await routeGuide.getFeature(point)
-
-      if feature.name.isEmpty {
-        print("No feature found at (\(self.latitude), \(self.longitude))")
-      } else {
-        print("Found '\(feature.name)' at (\(self.latitude), \(self.longitude))")
+        for (message, (lat, lon)) in notes {
+          let note = Routeguide_RouteNote.with {
+            $0.message = message
+            $0.location.latitude = lat
+            $0.location.longitude = lon
+          }
+          print("Sending note: '\(message) at (\(lat), \(lon))'")
+          try await writer.write(note)
+        }
+      } onResponse: { response in
+        for try await note in response.messages {
+          let (lat, lon) = (note.location.latitude, note.location.longitude)
+          print("Received note: '\(note.message) at (\(lat), \(lon))'")
+        }
       }
 
       client.beginGracefulShutdown()
