--- Examples/v2/hello-world/HelloWorld.proto
@@ -1,1 +1,0 @@
-../../../Protos/upstream/grpc/examples/helloworld.proto
